Remove ErrWebhookValidation (see #52)
Sparkpost's special "validation" call on webhook receivers should not be
treated as an error. It's cleaner to treat this as a normal webhook call
with zero number of events. This way, the users of the package don't
have to treat this single error separately.

This commit does just that: it removes ErrWebhookValidation to return an
empty list of events and a nil error for validation requests. Now, if
parsing webhook data returns an error it means that there really is an
error. No need for special "if"s to ignore ErrWebhookValidation.